image_captioning_backend/requirements.txt
Flask
requests
Pillow
flask_cors
gunicorn == 20.0.4